+++ src/components/Nav.tsx
@@ -1,4 +1,4 @@
-import { ACCESS_TOKEN_KEY } from '@/constants/token.contant';
+import { ACCESS_TOKEN_KEY } from '@/constants/token.constant';
 import routerMeta, { IRouterMeta } from '@/lib/routerMeta';
 import token from '@/lib/token';
 import { NavLink, useNavigate } from 'react-router-dom';
