+++ src/contexts/UserContextProvider.tsx
@@ -1,4 +1,4 @@
-import { ACCESS_TOKEN_KEY } from '@/constants/token.contant';
+import { ACCESS_TOKEN_KEY } from '@/constants/token.constant';
 import token from '@/lib/token';
 import { createContext, useState } from 'react';
